move msga into its own subcommand
